requirements.txt
# Core dependencies
python-dotenv>=1.0.1

# LangChain ecosystem (LangChain v1 with create_agent)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.39

# Other agent frameworks (optional - for comparison)
# git+https://github.com/openai/swarm.git@main#egg=swarm
# autogen~=0.2
# crewai
# crewai-tools

# Utilities
colorama>=0.4.0
graphviz>=0.20.0
pydantic>=2.0.0

# Evaluation framework dependencies
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
psutil>=5.8.0
numpy>=1.21.0

# Optional: Web UI
# gradio==5.11.0
